requirements.txt
transformers==4.41.2
torch==2.2.2
numpy<2
urllib3<2
accelerate==0.30.1
scipy==1.12.0
tokenizers==0.19.1
gradio==3.50.2
pymupdf
python-docx
PyPDF2
langdetect
deep-translator
matplotlib
docx2txt